fix: components structuring
