+++ client/src/components/Company/CompanyCard.jsx
