+++ client/src/components/index.js
@@ -1,14 +1,14 @@
 import Navbar from "./Navbar";
 import Footer from "./Footer";
 import CustomButton from "./CustomButton";
-import TextInput from "./TextInput";
 import SignUp from "./SignUp";
 import Header from "./Header";
 import ListBox from "./ListBox";
-import JobCard from "./JobCard";
 import Loading from "./Loading";
-import CompanyCard from "./CompanyCard";
-import JobTypes from "./JobTypes";
+import CompanyCard from "./Company/CompanyCard";
+import JobCard from "./Job/Card";
+import JobTypes from "./Job/Types";
+import TextInput from "./TextInput";
 
 export {
   Navbar,
@@ -18,8 +18,8 @@ export {
   SignUp,
   Header,
   ListBox,
-  JobCard,
+  JobTypes,
   Loading,
   CompanyCard,
-  JobTypes,
+  JobCard,
 };
